genaipoc1/health-checker/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
kubernetes==28.1.0
